requirements.txt
# Math libraries for machine learning and more
torch 
numpy
# Chess library to load and parse chess games
python-chess
# Visualization and plotting
tqdm
matplotlib
cairosvg

# Local library to import helpful functions and the trained model
-e .